clase tipo usuario

+++ index/src/main/java/crud/index/repositorio/Repository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import crud.index.modelos.Estado;
 
-public interface Repository extends JpaRepository<Estado, Long> {
+public interface Repository extends JpaRepository<Estado, Integer> {
+
+ 
 
 }
